Binding Update Added

+++ ERP_V2/src/main/java/com/example/ERP_V2/Services/BindingService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface BindingService {
     List<Binding> getAllBindings();
     void createBinding(Binding binding);
-    void updateBinding(Long id, Binding updatedBinding);
+    void updateBinding(int id, Binding updatedBinding);
 }
